+++ src/heat.js
@@ -484,6 +484,10 @@
                     createElementWithHTML( days, "div", "day-name", _configuration.dayNames[ dayNameIndex ] );
                 }
             }
+
+            if ( bindingOptions.views.map.showDaysInReverseOrder ) {
+                reverseElementsOrder( days );
+            }
         }
 
         var months = createElement( map, "div", "months" ),
@@ -525,6 +529,10 @@
                         }
         
                         if ( ( dayIndex + 1 ) % 7 === 0 ) {
+                            if ( bindingOptions.views.map.showDaysInReverseOrder ) {
+                                reverseElementsOrder( currentDayColumn );
+                            }
+
                             currentDayColumn = createElement( dayColumns, "div", "day-column" );
                             actualDay = 0;
 
@@ -1348,6 +1356,7 @@
         options.views.map.placeMonthNamesOnTheBottom = getDefaultBoolean( options.views.map.placeMonthNamesOnTheBottom, false );
         options.views.map.showDayNumbers = getDefaultBoolean( options.views.map.showDayNumbers, false );
         options.views.map.showMonthNames = getDefaultBoolean( options.views.map.showMonthNames, true );
+        options.views.map.showDaysInReverseOrder = getDefaultBoolean( options.views.map.showDaysInReverseOrder, false );
 
         if ( isInvalidOptionArray( options.views.map.monthsToShow ) ) {
             options.views.map.monthsToShow = _default_MonthsToShow;
@@ -1661,6 +1670,15 @@
         
         element.style.left = left + "px";
         element.style.top = top + "px";
+    }
+
+    function reverseElementsOrder( parent ) {
+        var children = parent.children,
+            childrenLength = children.length - 1;
+
+        for ( ; childrenLength--; ) {
+            parent.appendChild( children[ childrenLength ] );
+        }
     }